InputColor3 and InputColor4 + SeparatorText
The InputColor3 and InputColor4, based off the drag slider are 'working' to an extent.
There are many bugs and text formatting issues, but a bunch of good changes. The issues
lies mostly in the fact that there are so many different options and customisations.
Ideally I would create an entirely different class, but I really don't want to because
it's more bloat. The solution is faily generic but still needs fixing. The other issue
lies between converting from floats to integers and back. This is a headache and has
caused a lot of pain.

The SeparatorText was much easier, comparatively.