Update ZIndex and LayoutOrder when late renders occur
Currently, Iris does not well support components which conditionally
render, or render later (see #58 and #48).

This patch marks parents as dirty when a new component is rendered,
which triggers an update of the computed ZIndex and LayoutOrder for
those future children of the parent. This correctly orders elements,
albeit with a ~1 cycle delay in the order looking correct (minor
flickering).

This patch is not completely ideal -- as we naively assume that setting
the LayoutOrder and ZIndex of the widget's root Instance is sufficient.
This assumption breaks down as Iris uses ZIndexBehavior=Global, and so
we need to propagate an update of ZIndex/LayoutOrder to all instances
generated by a widget -- not just the root widget.

For simple cases however, this patch is suitable.