Changed GuiInset behaviour
There exists a mismatch between the coordinate system of the mouse position and that used in the AbsolutePosition and the root frame position. Initially, this was only an issue for the topbar inset but this is also an issue when the root frame is not positioned at the (0, 0) coordinate position for either mouse or AbsolutePosition coordinate systems.

To fix this, two offsets are used, one for the mouse and one for the absolute position:
- MouseOffset - which is used to only modify the mouse position
- GuiOffset - which is used to modify every AbsolutePosition

Outside of a play session, where there is not a topbar, both should be the same and be the absolute position of the root frame position. This is necessary for stories or plugins where the root Iris frame does not cover the entire screen.

In a regular play session where the topbar exists, they will be different depending on the IgnoreGuiInset property:
- false - GuiOffset is (0, 0) and MouseOffset is (0, -58) because the AbsolutePosition and Root frame are the same but the mouse position is offset by the topbar.
- true - GuiOffset is (0, -58) and MouseOffset is (0, 0) because the Root frame and mouse position have the same coordinates but the AbsolutePosition is offset by the topbar.