Fixed late update bug.
Fixed a bug where a widget would wait a frame to update but would not check whether it had been destroyed. Most promimently causing issues when the style of a slider widget constantly changed and was therefore destroyed and created every frame. Fixed by adding a check to make sure the widget was not discarded.