Added new State constructors.
Two new state constructors for working with local variables and table entries, design to bridge between luau data types and the state object. VariableState takes a variable and will sync the state and variable. TableState takes a table and key and will sync the state and value. Also added optional argument to disable the Iris cycle.